Remove color change with time setting, add rand base bounds
Remove change of yellow point when setting time, when the base was above 20 (or below -20) digit color is part of the number so the idea of hours = red and minutes = green doesn't hold. Also the visual change was more jarring than helpful when setting time. It may be worth adding back in color shift only when overlap mode is enabled, further use and testing will be needed.

Added bounding to the random base as bases on the extreme end of the range tend to be less interesting to look at as there isn't much difference in the display of time in such bases looks very similar for most times as the differences only appear when the seconds/minutes are very high.